refactor: clean up test duplication and fix useless assertions
- Fix test_compact_trigger_deserialization and test_pre_compact_payload_deserialization
  to wrap matches! in assert! (previously results were ignored)
- Remove duplicate extract_bash_commands tests from src/config_test.rs
- Remove duplicate HookResult constructor tests from src/types.rs
- Consolidate SubagentStop/SubagentStart payload validation tests into tests/types_tests.rs
- Remove overlapping extract_file_path, is_root_addition, matches_uneditable_pattern,
  and matches_agent_pattern tests from src/hooks_tests.rs and src/hooks.rs
- Remove low-value glob::Pattern behavior tests (tests external crate, not our logic)
- Remove pure serde round-trip tests for StopCommand/StopConfig
- Remove get_schema_url constant string assertion test
- Delete tests/settings_parsing_test.rs (replica struct serde tests)

Net reduction of ~1500 lines while maintaining comprehensive coverage.
