+++ src/hooks_tests.rs
@@ -73,94 +73,6 @@ fn test_notification_flag_gating_logic() {
     assert!(config_system_only.is_enabled_for("SessionStart"));
 }
 
-#[test]
-fn test_is_root_addition() {
-    use std::env;
-
-    // Get current working directory for testing
-    let cwd = env::current_dir().unwrap();
-
-    // Simulate config file in the current directory
-    let config_path = cwd.join(".conclaude.yaml");
-
-    // Files at the same level as config should be blocked
-    assert!(is_root_addition("test.txt", "test.txt", &config_path));
-    assert!(is_root_addition("newfile.rs", "newfile.rs", &config_path));
-
-    // BREAKING CHANGE: Dotfiles are now also blocked at root level
-    assert!(is_root_addition(".gitignore", ".gitignore", &config_path));
-    assert!(is_root_addition(".env", ".env", &config_path));
-
-    // BREAKING CHANGE: Config files are now also blocked at root level
-    assert!(is_root_addition(
-        "package.json",
-        "package.json",
-        &config_path
-    ));
-    assert!(is_root_addition("config.yaml", "config.yaml", &config_path));
-
-    // Files in subdirectories should not be blocked
-    assert!(!is_root_addition(
-        "src/test.txt",
-        "src/test.txt",
-        &config_path
-    ));
-    assert!(!is_root_addition(
-        "tests/foo.rs",
-        "tests/foo.rs",
-        &config_path
-    ));
-    assert!(!is_root_addition(
-        "nested/deep/file.txt",
-        "nested/deep/file.txt",
-        &config_path
-    ));
-}
-
-#[test]
-fn test_matches_uneditable_pattern() {
-    assert!(matches_uneditable_pattern(
-        "package.json",
-        "package.json",
-        "/path/package.json",
-        "package.json"
-    )
-    .unwrap());
-    assert!(matches_uneditable_pattern("test.md", "test.md", "/path/test.md", "*.md").unwrap());
-    assert!(matches_uneditable_pattern(
-        "src/index.ts",
-        "src/index.ts",
-        "/path/src/index.ts",
-        "src/**/*.ts"
-    )
-    .unwrap());
-    assert!(
-        !matches_uneditable_pattern("other.txt", "other.txt", "/path/other.txt", "*.md").unwrap()
-    );
-}
-
-#[test]
-fn test_extract_file_path() {
-    let mut tool_input = std::collections::HashMap::new();
-    tool_input.insert(
-        "file_path".to_string(),
-        Value::String("test.txt".to_string()),
-    );
-    assert_eq!(extract_file_path(&tool_input), Some("test.txt".to_string()));
-
-    tool_input.clear();
-    tool_input.insert(
-        "notebook_path".to_string(),
-        Value::String("notebook.ipynb".to_string()),
-    );
-    assert_eq!(
-        extract_file_path(&tool_input),
-        Some("notebook.ipynb".to_string())
-    );
-
-    tool_input.clear();
-    assert_eq!(extract_file_path(&tool_input), None);
-}
 
 #[test]
 fn test_truncate_output_no_truncation_needed() {
@@ -469,67 +381,6 @@ fn test_bash_command_wildcard_variations() {
 }
 
 // Tests for subagent stop pattern matching
-#[test]
-fn test_match_subagent_patterns_wildcard() {
-    use crate::config::{SubagentStopCommand, SubagentStopConfig};
-
-    let mut commands = std::collections::HashMap::new();
-    commands.insert(
-        "*".to_string(),
-        vec![SubagentStopCommand {
-            run: "echo all".to_string(),
-            message: None,
-            show_command: None,
-            show_stdout: None,
-            show_stderr: None,
-            max_output_lines: None,
-            timeout: None,
-        }],
-    );
-
-    let config = SubagentStopConfig { commands };
-
-    let matches = match_subagent_patterns("coder", &config).unwrap();
-    assert_eq!(matches, vec!["*"]);
-
-    let matches = match_subagent_patterns("tester", &config).unwrap();
-    assert_eq!(matches, vec!["*"]);
-
-    let matches = match_subagent_patterns("any-agent-name", &config).unwrap();
-    assert_eq!(matches, vec!["*"]);
-}
-
-#[test]
-fn test_match_subagent_patterns_exact() {
-    use crate::config::{SubagentStopCommand, SubagentStopConfig};
-
-    let mut commands = std::collections::HashMap::new();
-    commands.insert(
-        "coder".to_string(),
-        vec![SubagentStopCommand {
-            run: "echo coder".to_string(),
-            message: None,
-            show_command: None,
-            show_stdout: None,
-            show_stderr: None,
-            max_output_lines: None,
-            timeout: None,
-        }],
-    );
-
-    let config = SubagentStopConfig { commands };
-
-    let matches = match_subagent_patterns("coder", &config).unwrap();
-    assert_eq!(matches, vec!["coder"]);
-
-    // Should NOT match auto-coder or coder-agent
-    let matches = match_subagent_patterns("auto-coder", &config).unwrap();
-    assert!(matches.is_empty());
-
-    let matches = match_subagent_patterns("coder-agent", &config).unwrap();
-    assert!(matches.is_empty());
-}
-
 #[test]
 fn test_match_subagent_patterns_prefix_glob() {
     use crate::config::{SubagentStopCommand, SubagentStopConfig};
