+++ src/types.rs
@@ -291,22 +291,6 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_hook_result_success() {
-        let result = HookResult::success();
-        assert_eq!(result.blocked, Some(false));
-        assert_eq!(result.message, None);
-        assert_eq!(result.system_prompt, None);
-    }
-
-    #[test]
-    fn test_hook_result_blocked() {
-        let result = HookResult::blocked("Test blocking message");
-        assert_eq!(result.blocked, Some(true));
-        assert_eq!(result.message, Some("Test blocking message".to_string()));
-        assert_eq!(result.system_prompt, None);
-    }
-
-    #[test]
     fn test_hook_result_with_context() {
         let result = HookResult::with_context("Read the sidebar documentation");
         assert_eq!(result.blocked, Some(false));
@@ -338,155 +322,6 @@ mod tests {
         assert!(validate_base_payload(&invalid_base).is_err());
     }
 
-    #[test]
-    fn test_validate_subagent_stop_payload_valid() {
-        let payload = SubagentStopPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStop".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            stop_hook_active: true,
-            agent_id: "coder".to_string(),
-            agent_transcript_path: "/path/to/agent/transcript".to_string(),
-        };
-        assert!(validate_subagent_stop_payload(&payload).is_ok());
-    }
-
-    #[test]
-    fn test_validate_subagent_stop_payload_empty_agent_id() {
-        let payload = SubagentStopPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStop".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            stop_hook_active: true,
-            agent_id: String::new(),
-            agent_transcript_path: "/path/to/agent/transcript".to_string(),
-        };
-        let result = validate_subagent_stop_payload(&payload);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "agent_id cannot be empty");
-    }
-
-    #[test]
-    fn test_validate_subagent_stop_payload_whitespace_agent_id() {
-        let payload = SubagentStopPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStop".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            stop_hook_active: true,
-            agent_id: "   ".to_string(),
-            agent_transcript_path: "/path/to/agent/transcript".to_string(),
-        };
-        let result = validate_subagent_stop_payload(&payload);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "agent_id cannot be empty");
-    }
-
-    #[test]
-    fn test_validate_subagent_stop_payload_empty_agent_transcript_path() {
-        let payload = SubagentStopPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStop".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            stop_hook_active: true,
-            agent_id: "coder".to_string(),
-            agent_transcript_path: String::new(),
-        };
-        let result = validate_subagent_stop_payload(&payload);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "agent_transcript_path cannot be empty");
-    }
-
-    #[test]
-    fn test_validate_subagent_stop_payload_whitespace_agent_transcript_path() {
-        let payload = SubagentStopPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStop".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            stop_hook_active: true,
-            agent_id: "coder".to_string(),
-            agent_transcript_path: "   ".to_string(),
-        };
-        let result = validate_subagent_stop_payload(&payload);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "agent_transcript_path cannot be empty");
-    }
-
-    #[test]
-    fn test_validate_subagent_stop_payload_invalid_base() {
-        let payload = SubagentStopPayload {
-            base: BasePayload {
-                session_id: String::new(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStop".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            stop_hook_active: true,
-            agent_id: "coder".to_string(),
-            agent_transcript_path: "/path/to/agent/transcript".to_string(),
-        };
-        let result = validate_subagent_stop_payload(&payload);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("session_id"));
-    }
-
-    #[test]
-    fn test_validate_subagent_stop_payload_agent_id_with_leading_trailing_spaces() {
-        let payload = SubagentStopPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStop".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            stop_hook_active: true,
-            agent_id: "  coder  ".to_string(),
-            agent_transcript_path: "/path/to/agent/transcript".to_string(),
-        };
-        assert!(validate_subagent_stop_payload(&payload).is_ok());
-    }
-
-    #[test]
-    fn test_validate_subagent_stop_payload_different_agent_types() {
-        let agent_types = vec!["coder", "tester", "stuck"];
-
-        for agent_type in agent_types {
-            let payload = SubagentStopPayload {
-                base: BasePayload {
-                    session_id: "test_session".to_string(),
-                    transcript_path: "/path/to/transcript".to_string(),
-                    hook_event_name: "SubagentStop".to_string(),
-                    cwd: "/current/dir".to_string(),
-                    permission_mode: Some("default".to_string()),
-                },
-                stop_hook_active: true,
-                agent_id: agent_type.to_string(),
-                agent_transcript_path: "/path/to/agent/transcript".to_string(),
-            };
-            assert!(validate_subagent_stop_payload(&payload).is_ok());
-        }
-    }
 
     #[test]
     fn test_pre_tool_use_payload_serialization_with_tool_use_id() {
@@ -614,262 +449,6 @@ mod tests {
             deserialized.tool_use_id,
             Some("round-trip-id-2".to_string())
         );
-    }
-
-    #[test]
-    fn test_validate_subagent_start_payload_valid() {
-        let payload = SubagentStartPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStart".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            agent_id: "coder".to_string(),
-            subagent_type: "implementation".to_string(),
-            agent_transcript_path: "/path/to/agent/transcript".to_string(),
-        };
-        assert!(validate_subagent_start_payload(&payload).is_ok());
-    }
-
-    #[test]
-    fn test_validate_subagent_start_payload_empty_agent_id() {
-        let payload = SubagentStartPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStart".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            agent_id: String::new(),
-            subagent_type: "implementation".to_string(),
-            agent_transcript_path: "/path/to/agent/transcript".to_string(),
-        };
-        let result = validate_subagent_start_payload(&payload);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "agent_id cannot be empty");
-    }
-
-    #[test]
-    fn test_validate_subagent_start_payload_whitespace_agent_id() {
-        let payload = SubagentStartPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStart".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            agent_id: "   ".to_string(),
-            subagent_type: "implementation".to_string(),
-            agent_transcript_path: "/path/to/agent/transcript".to_string(),
-        };
-        let result = validate_subagent_start_payload(&payload);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "agent_id cannot be empty");
-    }
-
-    #[test]
-    fn test_validate_subagent_start_payload_empty_subagent_type() {
-        let payload = SubagentStartPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStart".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            agent_id: "coder".to_string(),
-            subagent_type: String::new(),
-            agent_transcript_path: "/path/to/agent/transcript".to_string(),
-        };
-        let result = validate_subagent_start_payload(&payload);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "subagent_type cannot be empty");
-    }
-
-    #[test]
-    fn test_validate_subagent_start_payload_whitespace_subagent_type() {
-        let payload = SubagentStartPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStart".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            agent_id: "coder".to_string(),
-            subagent_type: "   ".to_string(),
-            agent_transcript_path: "/path/to/agent/transcript".to_string(),
-        };
-        let result = validate_subagent_start_payload(&payload);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "subagent_type cannot be empty");
-    }
-
-    #[test]
-    fn test_validate_subagent_start_payload_empty_agent_transcript_path() {
-        let payload = SubagentStartPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStart".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            agent_id: "coder".to_string(),
-            subagent_type: "implementation".to_string(),
-            agent_transcript_path: String::new(),
-        };
-        let result = validate_subagent_start_payload(&payload);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "agent_transcript_path cannot be empty");
-    }
-
-    #[test]
-    fn test_validate_subagent_start_payload_whitespace_agent_transcript_path() {
-        let payload = SubagentStartPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStart".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            agent_id: "coder".to_string(),
-            subagent_type: "implementation".to_string(),
-            agent_transcript_path: "   ".to_string(),
-        };
-        let result = validate_subagent_start_payload(&payload);
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "agent_transcript_path cannot be empty");
-    }
-
-    #[test]
-    fn test_validate_subagent_start_payload_invalid_base() {
-        let payload = SubagentStartPayload {
-            base: BasePayload {
-                session_id: String::new(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStart".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            agent_id: "coder".to_string(),
-            subagent_type: "implementation".to_string(),
-            agent_transcript_path: "/path/to/agent/transcript".to_string(),
-        };
-        let result = validate_subagent_start_payload(&payload);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("session_id"));
-    }
-
-    #[test]
-    fn test_validate_subagent_start_payload_agent_id_with_leading_trailing_spaces() {
-        let payload = SubagentStartPayload {
-            base: BasePayload {
-                session_id: "test_session".to_string(),
-                transcript_path: "/path/to/transcript".to_string(),
-                hook_event_name: "SubagentStart".to_string(),
-                cwd: "/current/dir".to_string(),
-                permission_mode: Some("default".to_string()),
-            },
-            agent_id: "  coder  ".to_string(),
-            subagent_type: "implementation".to_string(),
-            agent_transcript_path: "/path/to/agent/transcript".to_string(),
-        };
-        assert!(validate_subagent_start_payload(&payload).is_ok());
-    }
-
-    #[test]
-    fn test_validate_subagent_start_payload_different_agent_types() {
-        let agent_types = vec!["coder", "tester", "stuck"];
-
-        for agent_type in agent_types {
-            let payload = SubagentStartPayload {
-                base: BasePayload {
-                    session_id: "test_session".to_string(),
-                    transcript_path: "/path/to/transcript".to_string(),
-                    hook_event_name: "SubagentStart".to_string(),
-                    cwd: "/current/dir".to_string(),
-                    permission_mode: Some("default".to_string()),
-                },
-                agent_id: agent_type.to_string(),
-                subagent_type: "implementation".to_string(),
-                agent_transcript_path: "/path/to/agent/transcript".to_string(),
-            };
-            assert!(validate_subagent_start_payload(&payload).is_ok());
-        }
-    }
-
-    #[test]
-    fn test_subagent_start_payload_deserialize_from_valid_json() {
-        let json = r#"{
-            "session_id": "test_session",
-            "transcript_path": "/path/to/transcript",
-            "hook_event_name": "SubagentStart",
-            "cwd": "/current/dir",
-            "permission_mode": "default",
-            "agent_id": "coder",
-            "subagent_type": "implementation",
-            "agent_transcript_path": "/path/to/agent/transcript"
-        }"#;
-
-        let payload: SubagentStartPayload = serde_json::from_str(json).unwrap();
-        assert_eq!(payload.agent_id, "coder");
-        assert_eq!(payload.subagent_type, "implementation");
-        assert_eq!(payload.agent_transcript_path, "/path/to/agent/transcript");
-        assert_eq!(payload.base.session_id, "test_session");
-    }
-
-    #[test]
-    fn test_subagent_start_payload_deserialize_missing_agent_id() {
-        let json = r#"{
-            "session_id": "test_session",
-            "transcript_path": "/path/to/transcript",
-            "hook_event_name": "SubagentStart",
-            "cwd": "/current/dir",
-            "permission_mode": "default",
-            "subagent_type": "implementation",
-            "agent_transcript_path": "/path/to/agent/transcript"
-        }"#;
-
-        let result: Result<SubagentStartPayload, _> = serde_json::from_str(json);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_subagent_start_payload_deserialize_missing_subagent_type() {
-        let json = r#"{
-            "session_id": "test_session",
-            "transcript_path": "/path/to/transcript",
-            "hook_event_name": "SubagentStart",
-            "cwd": "/current/dir",
-            "permission_mode": "default",
-            "agent_id": "coder",
-            "agent_transcript_path": "/path/to/agent/transcript"
-        }"#;
-
-        let result: Result<SubagentStartPayload, _> = serde_json::from_str(json);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_subagent_start_payload_deserialize_missing_agent_transcript_path() {
-        let json = r#"{
-            "session_id": "test_session",
-            "transcript_path": "/path/to/transcript",
-            "hook_event_name": "SubagentStart",
-            "cwd": "/current/dir",
-            "permission_mode": "default",
-            "agent_id": "coder",
-            "subagent_type": "implementation"
-        }"#;
-
-        let result: Result<SubagentStartPayload, _> = serde_json::from_str(json);
-        assert!(result.is_err());
     }
 
     // Tests for context injection - verifying HookResult behavior
