+++ tests/hooks_tests.rs
@@ -930,119 +930,6 @@ fn test_subagent_stop_payload_serialization() {
 }
 
 #[test]
-fn test_validate_subagent_stop_payload_all_fields_valid() {
-    let payload = create_subagent_stop_payload();
-    let result = validate_subagent_stop_payload(&payload);
-
-    assert!(result.is_ok(), "Valid payload should pass validation");
-}
-
-#[test]
-fn test_validate_subagent_stop_payload_missing_base_session_id() {
-    let mut payload = create_subagent_stop_payload();
-    payload.base.session_id = String::new();
-
-    let result = validate_subagent_stop_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with empty session_id should fail validation"
-    );
-    let error = result.unwrap_err();
-    assert!(
-        error.contains("session_id") || error.contains("Missing required field"),
-        "Error should mention session_id or missing field: {}",
-        error
-    );
-}
-
-#[test]
-fn test_validate_subagent_stop_payload_missing_base_transcript_path() {
-    let mut payload = create_subagent_stop_payload();
-    payload.base.transcript_path = String::new();
-
-    let result = validate_subagent_stop_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with empty transcript_path should fail validation"
-    );
-    let error = result.unwrap_err();
-    assert!(
-        error.contains("transcript_path") || error.contains("Missing required field"),
-        "Error should mention transcript_path or missing field: {}",
-        error
-    );
-}
-
-#[test]
-fn test_validate_subagent_stop_payload_missing_agent_id() {
-    let mut payload = create_subagent_stop_payload();
-    payload.agent_id = String::new();
-
-    let result = validate_subagent_stop_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with empty agent_id should fail validation"
-    );
-    assert!(result.unwrap_err().contains("agent_id cannot be empty"));
-}
-
-#[test]
-fn test_validate_subagent_stop_payload_whitespace_agent_id() {
-    let mut payload = create_subagent_stop_payload();
-    payload.agent_id = "   ".to_string(); // Only whitespace
-
-    let result = validate_subagent_stop_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with whitespace-only agent_id should fail validation"
-    );
-    assert!(result.unwrap_err().contains("agent_id cannot be empty"));
-}
-
-#[test]
-fn test_validate_subagent_stop_payload_missing_agent_transcript_path() {
-    let mut payload = create_subagent_stop_payload();
-    payload.agent_transcript_path = String::new();
-
-    let result = validate_subagent_stop_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with empty agent_transcript_path should fail validation"
-    );
-    assert!(result
-        .unwrap_err()
-        .contains("agent_transcript_path cannot be empty"));
-}
-
-#[test]
-fn test_validate_subagent_stop_payload_whitespace_agent_transcript_path() {
-    let mut payload = create_subagent_stop_payload();
-    payload.agent_transcript_path = "   ".to_string(); // Only whitespace
-
-    let result = validate_subagent_stop_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with whitespace-only agent_transcript_path should fail validation"
-    );
-    assert!(result
-        .unwrap_err()
-        .contains("agent_transcript_path cannot be empty"));
-}
-
-#[test]
-fn test_validate_subagent_stop_payload_multiple_missing_fields() {
-    let mut payload = create_subagent_stop_payload();
-    payload.agent_id = String::new();
-    payload.agent_transcript_path = String::new();
-
-    let result = validate_subagent_stop_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with multiple missing fields should fail"
-    );
-}
-
-#[test]
 fn test_subagent_stop_payload_json_round_trip_coder() {
     let json_str = r#"{
         "session_id": "test_session_coder",
@@ -1470,404 +1357,6 @@ fn test_subagent_stop_validation_error_messages_specific() {
         error.contains("agent_id"),
         "Error should mention agent_id for whitespace-only value"
     );
-}
-
-// ============================================================================
-// Integration Tests for SubagentStart Hook
-// ============================================================================
-
-// Helper function to create a SubagentStart payload for testing
-fn create_subagent_start_payload() -> SubagentStartPayload {
-    SubagentStartPayload {
-        base: BasePayload {
-            session_id: "test_session_789".to_string(),
-            transcript_path: "/tmp/session_transcript.jsonl".to_string(),
-            hook_event_name: "SubagentStart".to_string(),
-            cwd: "/home/user/project".to_string(),
-            permission_mode: Some("default".to_string()),
-        },
-        agent_id: "coder".to_string(),
-        subagent_type: "implementation".to_string(),
-        agent_transcript_path: "/tmp/coder_transcript.jsonl".to_string(),
-    }
-}
-
-#[test]
-fn test_subagent_start_payload_structure() {
-    let payload = create_subagent_start_payload();
-
-    // Verify all required fields are present and populated
-    assert_eq!(payload.base.session_id, "test_session_789");
-    assert_eq!(
-        payload.base.transcript_path,
-        "/tmp/session_transcript.jsonl"
-    );
-    assert_eq!(payload.base.hook_event_name, "SubagentStart");
-    assert_eq!(payload.base.cwd, "/home/user/project");
-    assert!(payload.base.permission_mode.is_some());
-    assert_eq!(payload.agent_id, "coder");
-    assert_eq!(payload.subagent_type, "implementation");
-    assert_eq!(payload.agent_transcript_path, "/tmp/coder_transcript.jsonl");
-}
-
-#[test]
-fn test_subagent_start_payload_with_different_agent_ids() {
-    // Test with different agent IDs: coder, tester, stuck
-    let agent_ids = vec!["coder", "tester", "stuck", "orchestrator"];
-
-    for agent_id in agent_ids {
-        let mut payload = create_subagent_start_payload();
-        payload.agent_id = agent_id.to_string();
-
-        assert_eq!(payload.agent_id, agent_id);
-        assert!(validate_subagent_start_payload(&payload).is_ok());
-    }
-}
-
-#[test]
-fn test_subagent_start_payload_serialization() {
-    let payload = create_subagent_start_payload();
-
-    // Serialize to JSON
-    let json_str = serde_json::to_string(&payload).expect("Failed to serialize");
-
-    // Deserialize back
-    let deserialized: SubagentStartPayload =
-        serde_json::from_str(&json_str).expect("Failed to deserialize");
-
-    // Verify round-trip preservation
-    assert_eq!(deserialized.base.session_id, payload.base.session_id);
-    assert_eq!(deserialized.agent_id, payload.agent_id);
-    assert_eq!(deserialized.subagent_type, payload.subagent_type);
-    assert_eq!(
-        deserialized.agent_transcript_path,
-        payload.agent_transcript_path
-    );
-}
-
-#[test]
-fn test_validate_subagent_start_payload_all_fields_valid() {
-    let payload = create_subagent_start_payload();
-    let result = validate_subagent_start_payload(&payload);
-
-    assert!(result.is_ok(), "Valid payload should pass validation");
-}
-
-#[test]
-fn test_validate_subagent_start_payload_missing_base_session_id() {
-    let mut payload = create_subagent_start_payload();
-    payload.base.session_id = String::new();
-
-    let result = validate_subagent_start_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with empty session_id should fail validation"
-    );
-    let error = result.unwrap_err();
-    assert!(
-        error.contains("session_id") || error.contains("Missing required field"),
-        "Error should mention session_id or missing field: {}",
-        error
-    );
-}
-
-#[test]
-fn test_validate_subagent_start_payload_missing_agent_id() {
-    let mut payload = create_subagent_start_payload();
-    payload.agent_id = String::new();
-
-    let result = validate_subagent_start_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with empty agent_id should fail validation"
-    );
-    assert!(result.unwrap_err().contains("agent_id cannot be empty"));
-}
-
-#[test]
-fn test_validate_subagent_start_payload_whitespace_agent_id() {
-    let mut payload = create_subagent_start_payload();
-    payload.agent_id = "   ".to_string(); // Only whitespace
-
-    let result = validate_subagent_start_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with whitespace-only agent_id should fail validation"
-    );
-    assert!(result.unwrap_err().contains("agent_id cannot be empty"));
-}
-
-#[test]
-fn test_validate_subagent_start_payload_missing_subagent_type() {
-    let mut payload = create_subagent_start_payload();
-    payload.subagent_type = String::new();
-
-    let result = validate_subagent_start_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with empty subagent_type should fail validation"
-    );
-    assert!(result
-        .unwrap_err()
-        .contains("subagent_type cannot be empty"));
-}
-
-#[test]
-fn test_validate_subagent_start_payload_whitespace_subagent_type() {
-    let mut payload = create_subagent_start_payload();
-    payload.subagent_type = "   ".to_string(); // Only whitespace
-
-    let result = validate_subagent_start_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with whitespace-only subagent_type should fail validation"
-    );
-    assert!(result
-        .unwrap_err()
-        .contains("subagent_type cannot be empty"));
-}
-
-#[test]
-fn test_validate_subagent_start_payload_missing_agent_transcript_path() {
-    let mut payload = create_subagent_start_payload();
-    payload.agent_transcript_path = String::new();
-
-    let result = validate_subagent_start_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with empty agent_transcript_path should fail validation"
-    );
-    assert!(result
-        .unwrap_err()
-        .contains("agent_transcript_path cannot be empty"));
-}
-
-#[test]
-fn test_validate_subagent_start_payload_whitespace_agent_transcript_path() {
-    let mut payload = create_subagent_start_payload();
-    payload.agent_transcript_path = "   ".to_string(); // Only whitespace
-
-    let result = validate_subagent_start_payload(&payload);
-    assert!(
-        result.is_err(),
-        "Payload with whitespace-only agent_transcript_path should fail validation"
-    );
-    assert!(result
-        .unwrap_err()
-        .contains("agent_transcript_path cannot be empty"));
-}
-
-#[test]
-fn test_validate_subagent_start_payload_agent_id_with_leading_trailing_spaces() {
-    let mut payload = create_subagent_start_payload();
-    payload.agent_id = "  coder  ".to_string();
-
-    // Should pass validation because we trim before checking
-    assert!(validate_subagent_start_payload(&payload).is_ok());
-}
-
-#[test]
-fn test_validate_subagent_start_payload_invalid_base() {
-    let mut payload = create_subagent_start_payload();
-    payload.base.session_id = String::new();
-
-    let result = validate_subagent_start_payload(&payload);
-    assert!(result.is_err());
-    assert!(result.unwrap_err().contains("session_id"));
-}
-
-#[test]
-fn test_subagent_start_payload_json_round_trip_coder() {
-    let json_str = r#"{
-        "session_id": "test_session_coder",
-        "transcript_path": "/tmp/session_123.jsonl",
-        "hook_event_name": "SubagentStart",
-        "cwd": "/home/user/project",
-        "permission_mode": "default",
-        "agent_id": "coder",
-        "subagent_type": "implementation",
-        "agent_transcript_path": "/tmp/coder_123.jsonl"
-    }"#;
-
-    let payload: SubagentStartPayload =
-        serde_json::from_str(json_str).expect("Failed to deserialize JSON");
-
-    assert_eq!(payload.agent_id, "coder");
-    assert_eq!(payload.subagent_type, "implementation");
-    assert_eq!(payload.agent_transcript_path, "/tmp/coder_123.jsonl");
-    assert_eq!(payload.base.session_id, "test_session_coder");
-    assert!(validate_subagent_start_payload(&payload).is_ok());
-}
-
-#[test]
-fn test_subagent_start_payload_json_round_trip_tester() {
-    let json_str = r#"{
-        "session_id": "test_session_tester",
-        "transcript_path": "/tmp/session_456.jsonl",
-        "hook_event_name": "SubagentStart",
-        "cwd": "/home/user/project",
-        "permission_mode": "default",
-        "agent_id": "tester",
-        "subagent_type": "testing",
-        "agent_transcript_path": "/tmp/tester_456.jsonl"
-    }"#;
-
-    let payload: SubagentStartPayload =
-        serde_json::from_str(json_str).expect("Failed to deserialize JSON");
-
-    assert_eq!(payload.agent_id, "tester");
-    assert_eq!(payload.subagent_type, "testing");
-    assert_eq!(payload.agent_transcript_path, "/tmp/tester_456.jsonl");
-    assert!(validate_subagent_start_payload(&payload).is_ok());
-}
-
-#[test]
-fn test_subagent_start_payload_json_round_trip_stuck() {
-    let json_str = r#"{
-        "session_id": "test_session_stuck",
-        "transcript_path": "/tmp/session_789.jsonl",
-        "hook_event_name": "SubagentStart",
-        "cwd": "/home/user/project",
-        "permission_mode": "restrict",
-        "agent_id": "stuck",
-        "subagent_type": "escalation",
-        "agent_transcript_path": "/tmp/stuck_789.jsonl"
-    }"#;
-
-    let payload: SubagentStartPayload =
-        serde_json::from_str(json_str).expect("Failed to deserialize JSON");
-
-    assert_eq!(payload.agent_id, "stuck");
-    assert_eq!(payload.subagent_type, "escalation");
-    assert_eq!(payload.agent_transcript_path, "/tmp/stuck_789.jsonl");
-    assert_eq!(payload.base.permission_mode, Some("restrict".to_string()));
-    assert!(validate_subagent_start_payload(&payload).is_ok());
-}
-
-#[test]
-fn test_subagent_start_payload_json_with_missing_agent_id_field() {
-    let json_str = r#"{
-        "session_id": "test_session",
-        "transcript_path": "/tmp/session.jsonl",
-        "hook_event_name": "SubagentStart",
-        "cwd": "/home/user/project",
-        "permission_mode": "default",
-        "subagent_type": "implementation",
-        "agent_transcript_path": "/tmp/agent.jsonl"
-    }"#;
-
-    // This should fail to deserialize because agent_id is required
-    let result: Result<SubagentStartPayload, _> = serde_json::from_str(json_str);
-    assert!(
-        result.is_err(),
-        "JSON missing agent_id should fail to deserialize"
-    );
-}
-
-#[test]
-fn test_subagent_start_payload_json_with_missing_subagent_type_field() {
-    let json_str = r#"{
-        "session_id": "test_session",
-        "transcript_path": "/tmp/session.jsonl",
-        "hook_event_name": "SubagentStart",
-        "cwd": "/home/user/project",
-        "permission_mode": "default",
-        "agent_id": "coder",
-        "agent_transcript_path": "/tmp/agent.jsonl"
-    }"#;
-
-    // This should fail to deserialize because subagent_type is required
-    let result: Result<SubagentStartPayload, _> = serde_json::from_str(json_str);
-    assert!(
-        result.is_err(),
-        "JSON missing subagent_type should fail to deserialize"
-    );
-}
-
-#[test]
-fn test_subagent_start_payload_json_with_missing_agent_transcript_path_field() {
-    let json_str = r#"{
-        "session_id": "test_session",
-        "transcript_path": "/tmp/session.jsonl",
-        "hook_event_name": "SubagentStart",
-        "cwd": "/home/user/project",
-        "permission_mode": "default",
-        "agent_id": "coder",
-        "subagent_type": "implementation"
-    }"#;
-
-    // This should fail to deserialize because agent_transcript_path is required
-    let result: Result<SubagentStartPayload, _> = serde_json::from_str(json_str);
-    assert!(
-        result.is_err(),
-        "JSON missing agent_transcript_path should fail to deserialize"
-    );
-}
-
-#[test]
-fn test_subagent_start_validation_error_messages_specific() {
-    // Test that error messages are specific and helpful
-
-    let mut payload = create_subagent_start_payload();
-
-    // Test agent_id error
-    payload.agent_id = String::new();
-    let error = validate_subagent_start_payload(&payload).unwrap_err();
-    assert!(
-        error.contains("agent_id"),
-        "Error should mention agent_id field"
-    );
-
-    // Test subagent_type error
-    let mut payload = create_subagent_start_payload();
-    payload.subagent_type = String::new();
-    let error = validate_subagent_start_payload(&payload).unwrap_err();
-    assert!(
-        error.contains("subagent_type"),
-        "Error should mention subagent_type field"
-    );
-
-    // Test agent_transcript_path error
-    let mut payload = create_subagent_start_payload();
-    payload.agent_transcript_path = String::new();
-    let error = validate_subagent_start_payload(&payload).unwrap_err();
-    assert!(
-        error.contains("agent_transcript_path"),
-        "Error should mention agent_transcript_path field"
-    );
-
-    // Test whitespace-only agent_id error
-    let mut payload = create_subagent_start_payload();
-    payload.agent_id = "   ".to_string();
-    let error = validate_subagent_start_payload(&payload).unwrap_err();
-    assert!(
-        error.contains("agent_id"),
-        "Error should mention agent_id for whitespace-only value"
-    );
-}
-
-#[test]
-fn test_subagent_start_payload_with_different_subagent_types() {
-    let subagent_types = vec!["implementation", "testing", "escalation", "analysis"];
-
-    for subagent_type in subagent_types {
-        let mut payload = create_subagent_start_payload();
-        payload.subagent_type = subagent_type.to_string();
-
-        assert_eq!(payload.subagent_type, subagent_type);
-        assert!(validate_subagent_start_payload(&payload).is_ok());
-    }
-}
-
-#[test]
-fn test_subagent_start_hook_event_name_validation() {
-    let payload = create_subagent_start_payload();
-
-    // Verify the hook event name is correctly set
-    assert_eq!(payload.base.hook_event_name, "SubagentStart");
-
-    // Verify the base payload validates correctly
-    assert!(validate_base_payload(&payload.base).is_ok());
 }
